some refact

+++ src/types.h
@@ -6,5 +6,5 @@
 using bool_vec = std::vector<bool>;
 using bool_mat = std::vector<bool_vec>;
 using byte = std::uint8_t;
-using monom_t = std::uint16_t;
-using monom_and_deg = std::pair<monom_t, int>;
+using monom_t = std::uint64_t;
+using monom_and_deg = std::pair<monom_t, unsigned>;
